refactoring/c++: move more code from cpp/mod.rs
