+++ macroslib/src/cpp/mod.rs
@@ -42,31 +42,24 @@ use proc_macro2::TokenStream;
 use rustc_hash::{FxHashMap, FxHashSet};
 use smol_str::SmolStr;
 use strum::IntoEnumIterator;
-use syn::Type;
 
 use crate::{
     cpp::{
         map_class_self_type::register_typemap_for_self_type,
         map_type::{do_map_type, map_type},
     },
-    error::{invalid_src_id_span, DiagnosticError, Result},
+    error::{DiagnosticError, Result},
     file_cache::FileWriteCache,
     source_registry::SourceId,
     typemap::{
         ast::{parse_ty_with_given_span, TypeName},
-        ty::{
-            ForeignConversationIntermediate, ForeignConversationRule, ForeignType, ForeignTypeS,
-            RustType,
-        },
+        ty::{ForeignConversationRule, ForeignType, ForeignTypeS, RustType},
         utils::{
             configure_ftype_rule, validate_cfg_options, ForeignMethodSignature, ForeignTypeInfoT,
         },
-        CItem, CItems, ForeignTypeInfo, TypeConvCode, TypeMapConvRuleInfo, FROM_VAR_TEMPLATE,
+        CItem, CItems, ForeignTypeInfo, TypeMapConvRuleInfo,
     },
-    types::{
-        ForeignEnumInfo, ForeignInterface, ForeignerClassInfo, ForeignerMethod, ItemToExpand,
-        MethodAccess, MethodVariant,
-    },
+    types::{ForeignerClassInfo, ForeignerMethod, ItemToExpand, MethodAccess},
     CppConfig, CppOptional, CppStrView, CppVariant, LanguageGenerator, SourceCode, TypeMap,
     WRITE_TO_MEM_FAILED_MSG,
 };
@@ -268,10 +261,10 @@ impl LanguageGenerator for CppConfig {
             }
             for item in items {
                 match item {
-                    ItemToExpand::Class(fclass) => generate(&mut ctx, &fclass)?,
-                    ItemToExpand::Enum(fenum) => generate_enum(&mut ctx, &fenum)?,
+                    ItemToExpand::Class(fclass) => fclass::generate(&mut ctx, &fclass)?,
+                    ItemToExpand::Enum(fenum) => fenum::generate_enum(&mut ctx, &fenum)?,
                     ItemToExpand::Interface(finterface) => {
-                        generate_interface(&mut ctx, &finterface)?
+                        finterface::generate_interface(&mut ctx, &finterface)?
                     }
                 }
             }
@@ -494,129 +487,6 @@ fn init(ctx: &mut CppContext, code: &[SourceCode]) -> Result<()> {
     for rule in not_merged_data {
         merge_rule(ctx, rule)?;
     }
-
-    Ok(())
-}
-
-fn generate(ctx: &mut CppContext, class: &ForeignerClassInfo) -> Result<()> {
-    debug!(
-        "generate: begin for {}, this_type_for_method {:?}",
-        class.name, class.self_desc
-    );
-    let has_methods = class.methods.iter().any(|m| match m.variant {
-        MethodVariant::Method(_) => true,
-        _ => false,
-    });
-    let has_constructor = class
-        .methods
-        .iter()
-        .any(|m| m.variant == MethodVariant::Constructor);
-
-    if has_methods && !has_constructor {
-        return Err(DiagnosticError::new(
-            class.src_id,
-            class.span(),
-            format!(
-                "namespace {}, class {}: has methods, but no constructor\n
-May be you need to use `private constructor = empty;` syntax?",
-                ctx.cfg.namespace_name, class.name
-            ),
-        ));
-    }
-
-    let mut m_sigs = fclass::find_suitable_foreign_types_for_methods(ctx, class)?;
-    let mut req_includes = cpp_code::cpp_list_required_includes(&mut m_sigs);
-    let my_self_cpp = format!("\"{}\"", cpp_code::cpp_header_name(class));
-    let my_self_c = format!("\"{}\"", cpp_code::c_header_name(class));
-    req_includes.retain(|el| *el != my_self_cpp && *el != my_self_c);
-    fclass::generate(ctx, class, &req_includes, &m_sigs)?;
-    Ok(())
-}
-
-fn generate_enum(ctx: &mut CppContext, fenum: &ForeignEnumInfo) -> Result<()> {
-    if (fenum.items.len() as u64) >= u64::from(u32::max_value()) {
-        return Err(DiagnosticError::new(
-            fenum.src_id,
-            fenum.span(),
-            "Too many items in enum",
-        ));
-    }
-
-    trace!("enum_ti: {}", fenum.name);
-    let enum_name = &fenum.name;
-    let enum_ti: Type = parse_ty_with_given_span(&enum_name.to_string(), fenum.name.span())
-        .map_err(|err| DiagnosticError::from_syn_err(fenum.src_id, err))?;
-    let enum_rty = ctx.conv_map.find_or_alloc_rust_type_that_implements(
-        &enum_ti,
-        &["SwigForeignEnum"],
-        fenum.src_id,
-    );
-
-    fenum::generate_c_code_for_enum(&ctx.cfg.output_dir, fenum)
-        .map_err(|err| DiagnosticError::new(fenum.src_id, fenum.span(), err))?;
-    fenum::generate_rust_trait_for_enum(ctx, fenum)?;
-
-    let u32_rty = ctx
-        .conv_map
-        .find_or_alloc_rust_type_no_src_id(&parse_type! { u32 });
-
-    let enum_ftype = ForeignTypeS {
-        name: TypeName::new(fenum.name.to_string(), (fenum.src_id, fenum.name.span())),
-        provides_by_module: vec![
-            format!("\"{}\"", cpp_code::cpp_header_name_for_enum(fenum)).into()
-        ],
-        into_from_rust: Some(ForeignConversationRule {
-            rust_ty: enum_rty.to_idx(),
-            intermediate: Some(ForeignConversationIntermediate {
-                input_to_output: false,
-                intermediate_ty: u32_rty.to_idx(),
-                conv_code: TypeConvCode::new(
-                    format!(
-                        "static_cast<{enum_name}>({var})",
-                        enum_name = fenum.name,
-                        var = FROM_VAR_TEMPLATE
-                    ),
-                    invalid_src_id_span(),
-                ),
-            }),
-        }),
-        from_into_rust: Some(ForeignConversationRule {
-            rust_ty: enum_rty.to_idx(),
-            intermediate: Some(ForeignConversationIntermediate {
-                input_to_output: false,
-                intermediate_ty: u32_rty.to_idx(),
-                conv_code: TypeConvCode::new(
-                    format!("static_cast<uint32_t>({})", FROM_VAR_TEMPLATE),
-                    invalid_src_id_span(),
-                ),
-            }),
-        }),
-        name_prefix: None,
-    };
-    ctx.conv_map.alloc_foreign_type(enum_ftype)?;
-    ctx.conv_map.register_exported_enum(fenum);
-    Ok(())
-}
-
-fn generate_interface(ctx: &mut CppContext, interface: &ForeignInterface) -> Result<()> {
-    let mut f_methods = finterface::find_suitable_ftypes_for_interace_methods(ctx, interface)?;
-    let req_includes = cpp_code::cpp_list_required_includes(&mut f_methods);
-    finterface::generate_for_interface(ctx, interface, &req_includes, &f_methods)
-        .map_err(|err| DiagnosticError::new(interface.src_id, interface.span(), err))?;
-    finterface::rust_code_generate_interface(ctx, interface, &f_methods)?;
-
-    let c_struct_name = format!("C_{}", interface.name);
-    let rust_struct_pointer = format!("*const {}", c_struct_name);
-    let rust_ty: Type = parse_ty_with_given_span(&rust_struct_pointer, interface.name.span())
-        .map_err(|err| DiagnosticError::from_syn_err(interface.src_id, err))?;
-    let c_struct_pointer = format!("const struct {} * const", c_struct_name);
-
-    let rust_ty = ctx.conv_map.find_or_alloc_rust_type_no_src_id(&rust_ty);
-
-    ctx.conv_map.add_foreign(
-        rust_ty,
-        TypeName::new(c_struct_pointer, interface.src_id_span()),
-    )?;
 
     Ok(())
 }
